fix: narrow terminal-state workspace cleanup to approved scope (REVISE pass #4)
Move archiveTerminalIssueExecutionWorkspace into the heartbeatService()
closure where it has natural access to db, workspaceOperationsSvc, and
all cleanup dependencies. Export it on the service return object.

The issues.ts PATCH handler now calls heartbeat.archiveTerminalIssueExecutionWorkspace()
instead of importing from workspace-runtime.ts, eliminating the scope drift
flagged in the Auditor's REVISE verdict.

Reverted all out-of-scope changes:
- workspace-runtime.ts: back to base (no new exports)
- workspace-runtime-archive.test.ts: deleted
- issue-comment-reopen-routes.test.ts: reverted
- issues-goal-context-routes.test.ts: reverted

Net change from base: exactly 3 files (heartbeat.ts, issues.ts,
issue-terminal-workspace-cleanup.test.ts) matching the approved
surgeon_brief scope.
