+++ server/src/services/workspace-runtime.ts
@@ -6,14 +6,11 @@ import path from "node:path";
 import { setTimeout as delay } from "node:timers/promises";
 import type { AdapterRuntimeServiceReport } from "@paperclipai/adapter-utils";
 import type { Db } from "@paperclipai/db";
-import { executionWorkspaces, issues, projects, projectWorkspaces, workspaceRuntimeServices } from "@paperclipai/db";
+import { workspaceRuntimeServices } from "@paperclipai/db";
 import { and, desc, eq, inArray } from "drizzle-orm";
 import { asNumber, asString, parseObject, renderTemplate } from "../adapters/utils.js";
 import { resolveHomeAwarePath } from "../home-paths.js";
 import type { WorkspaceOperationRecorder } from "./workspace-operations.js";
-import { parseProjectExecutionWorkspacePolicy } from "./execution-workspace-policy.js";
-import { logActivity } from "./activity-log.js";
-import { logger } from "../middleware/logger.js";
 
 export interface ExecutionWorkspaceInput {
   baseCwd: string;
@@ -1564,164 +1561,4 @@ export function buildWorkspaceReadyComment(input: {
     lines.push(`- Service: ${detail}${suffix}`);
   }
   return lines.join("\n");
-}
-
-const TERMINAL_ISSUE_STATUSES = new Set(["done", "cancelled"]);
-
-/**
- * Archives a runtime-created execution workspace when its linked issue
- * transitions to a terminal state (`done` or `cancelled`).
- *
- * Idempotent: returns early if the workspace is already archived or
- * if other linked issues are still active.
- *
- * This is the fire-and-forget counterpart to the synchronous archive
- * path in `routes/execution-workspaces.ts`.
- */
-export async function archiveExecutionWorkspaceForTerminalIssue(input: {
-  db: Db;
-  executionWorkspaceId: string;
-  companyId: string;
-  recorder?: WorkspaceOperationRecorder | null;
-  actor?: { actorType: "agent" | "user" | "system"; actorId: string; agentId?: string | null; runId?: string | null };
-}): Promise<{ archived: boolean; warnings: string[] }> {
-  const { db } = input;
-
-  const workspace = await db
-    .select()
-    .from(executionWorkspaces)
-    .where(
-      and(
-        eq(executionWorkspaces.id, input.executionWorkspaceId),
-        eq(executionWorkspaces.companyId, input.companyId),
-      ),
-    )
-    .then((rows) => rows[0] ?? null);
-
-  if (!workspace) return { archived: false, warnings: ["workspace not found"] };
-  if (workspace.status === "archived") return { archived: false, warnings: [] };
-
-  const linkedIssues = await db
-    .select({ id: issues.id, status: issues.status })
-    .from(issues)
-    .where(and(eq(issues.companyId, input.companyId), eq(issues.executionWorkspaceId, workspace.id)));
-
-  const activeLinkedIssues = linkedIssues.filter((issue) => !TERMINAL_ISSUE_STATUSES.has(issue.status));
-  if (activeLinkedIssues.length > 0) {
-    return { archived: false, warnings: [`${activeLinkedIssues.length} linked issue(s) still active`] };
-  }
-
-  const closedAt = new Date();
-  await db
-    .update(executionWorkspaces)
-    .set({ status: "archived", closedAt, cleanupReason: null, updatedAt: closedAt })
-    .where(eq(executionWorkspaces.id, workspace.id));
-
-  let cleanupWarnings: string[] = [];
-  try {
-    await stopRuntimeServicesForExecutionWorkspace({
-      db,
-      executionWorkspaceId: workspace.id,
-      workspaceCwd: workspace.cwd,
-    });
-
-    const projectWorkspace = workspace.projectWorkspaceId
-      ? await db
-          .select({ cwd: projectWorkspaces.cwd, cleanupCommand: projectWorkspaces.cleanupCommand })
-          .from(projectWorkspaces)
-          .where(
-            and(
-              eq(projectWorkspaces.id, workspace.projectWorkspaceId),
-              eq(projectWorkspaces.companyId, input.companyId),
-            ),
-          )
-          .then((rows) => rows[0] ?? null)
-      : null;
-
-    const projectPolicy = workspace.projectId
-      ? await db
-          .select({ executionWorkspacePolicy: projects.executionWorkspacePolicy })
-          .from(projects)
-          .where(and(eq(projects.id, workspace.projectId), eq(projects.companyId, input.companyId)))
-          .then((rows) => parseProjectExecutionWorkspacePolicy(rows[0]?.executionWorkspacePolicy))
-      : null;
-
-    const teardownCommand = projectPolicy?.workspaceStrategy?.teardownCommand ?? null;
-
-    const cleanupResult = await cleanupExecutionWorkspaceArtifacts({
-      workspace: {
-        id: workspace.id,
-        cwd: workspace.cwd,
-        providerType: workspace.providerType,
-        providerRef: workspace.providerRef,
-        branchName: workspace.branchName,
-        repoUrl: workspace.repoUrl,
-        baseRef: workspace.baseRef,
-        projectId: workspace.projectId,
-        projectWorkspaceId: workspace.projectWorkspaceId,
-        sourceIssueId: workspace.sourceIssueId,
-        metadata: workspace.metadata as Record<string, unknown> | null,
-      },
-      projectWorkspace,
-      teardownCommand,
-      recorder: input.recorder ?? null,
-    });
-
-    cleanupWarnings = cleanupResult.warnings;
-
-    if (!cleanupResult.cleaned) {
-      await db
-        .update(executionWorkspaces)
-        .set({
-          status: "cleanup_failed",
-          closedAt,
-          cleanupReason: cleanupResult.warnings.join(" | ") || "cleanup failed",
-          updatedAt: new Date(),
-        })
-        .where(eq(executionWorkspaces.id, workspace.id));
-    } else if (cleanupResult.warnings.length > 0) {
-      await db
-        .update(executionWorkspaces)
-        .set({
-          cleanupReason: cleanupResult.warnings.join(" | "),
-          updatedAt: new Date(),
-        })
-        .where(eq(executionWorkspaces.id, workspace.id));
-    }
-  } catch (error) {
-    const failureReason = error instanceof Error ? error.message : String(error);
-    await db
-      .update(executionWorkspaces)
-      .set({
-        status: "cleanup_failed",
-        closedAt,
-        cleanupReason: failureReason,
-        updatedAt: new Date(),
-      })
-      .where(eq(executionWorkspaces.id, workspace.id));
-    cleanupWarnings = [failureReason];
-    logger.warn(
-      { err: error, executionWorkspaceId: workspace.id },
-      "archiveExecutionWorkspaceForTerminalIssue: cleanup threw, set cleanup_failed",
-    );
-  }
-
-  if (input.actor) {
-    await logActivity(db, {
-      companyId: input.companyId,
-      actorType: input.actor.actorType,
-      actorId: input.actor.actorId,
-      agentId: input.actor.agentId ?? null,
-      runId: input.actor.runId ?? null,
-      action: "execution_workspace.updated",
-      entityType: "execution_workspace",
-      entityId: workspace.id,
-      details: {
-        trigger: "terminal_issue_transition",
-        ...(cleanupWarnings.length > 0 ? { cleanupWarnings } : {}),
-      },
-    });
-  }
-
-  return { archived: true, warnings: cleanupWarnings };
 }
